# oas/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/agents": {
      "get": {
        "operationId": "AgentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              }
            }
          }
        },
        "summary": "모든 Agent 조회",
        "tags": [
          "Agent"
        ]
      },
      "post": {
        "operationId": "AgentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        },
        "summary": "새로운 Agent 생성",
        "tags": [
          "Agent"
        ]
      }
    },
    "/agents/{id}": {
      "get": {
        "operationId": "AgentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "404": {
            "description": "Agent가 존재하지 않음"
          }
        },
        "summary": "특정 Agent 조회",
        "tags": [
          "Agent"
        ]
      },
      "put": {
        "operationId": "AgentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "404": {
            "description": "Agent가 존재하지 않음"
          }
        },
        "summary": "Agent 정보 업데이트",
        "tags": [
          "Agent"
        ]
      },
      "delete": {
        "operationId": "AgentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공"
          },
          "404": {
            "description": "Agent가 존재하지 않음"
          }
        },
        "summary": "Agent 삭제",
        "tags": [
          "Agent"
        ]
      }
    },
    "/mcps": {
      "get": {
        "operationId": "McpsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mcp"
                  }
                }
              }
            }
          }
        },
        "summary": "모든 Mcp 설정 조회",
        "tags": [
          "Mcp"
        ]
      },
      "post": {
        "operationId": "McpsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMcpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mcp"
                }
              }
            }
          }
        },
        "summary": "새로운 Mcp 설정 생성",
        "tags": [
          "Mcp"
        ]
      }
    },
    "/mcps/{id}": {
      "get": {
        "operationId": "McpsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mcp"
                }
              }
            }
          },
          "404": {
            "description": "Mcp 설정이 존재하지 않음"
          }
        },
        "summary": "특정 Mcp 설정 조회",
        "tags": [
          "Mcp"
        ]
      },
      "put": {
        "operationId": "McpsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mcp"
                }
              }
            }
          },
          "404": {
            "description": "Mcp 설정이 존재하지 않음"
          }
        },
        "summary": "Mcp 설정 업데이트",
        "tags": [
          "Mcp"
        ]
      },
      "delete": {
        "operationId": "McpsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공"
          },
          "404": {
            "description": "Mcp 설정이 존재하지 않음"
          }
        },
        "summary": "Mcp 설정 삭제",
        "tags": [
          "Mcp"
        ]
      }
    },
    "/schedules": {
      "get": {
        "operationId": "SchedulesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "summary": "모든 예약 실행 조회",
        "tags": [
          "Schedule"
        ]
      },
      "post": {
        "operationId": "SchedulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          }
        },
        "summary": "새로운 예약 실행 생성",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedules/{id}": {
      "get": {
        "operationId": "SchedulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          },
          "404": {
            "description": "예약 실행이 존재하지 않음"
          }
        },
        "summary": "특정 예약 실행 조회",
        "tags": [
          "Schedule"
        ]
      },
      "put": {
        "operationId": "SchedulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          },
          "404": {
            "description": "예약 실행이 존재하지 않음"
          }
        },
        "summary": "예약 실행 업데이트",
        "tags": [
          "Schedule"
        ]
      },
      "delete": {
        "operationId": "SchedulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공"
          },
          "404": {
            "description": "예약 실행이 존재하지 않음"
          }
        },
        "summary": "예약 실행 삭제",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedules/agent/{agentId}": {
      "get": {
        "operationId": "SchedulesController_findByAgentId",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "summary": "특정 Agent의 모든 예약 실행 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/chat": {
      "get": {
        "operationId": "ChatController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessage"
                  }
                }
              }
            }
          }
        },
        "summary": "모든 채팅 메시지 조회",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "operationId": "ChatController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          }
        },
        "summary": "새로운 채팅 메시지 생성",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/{id}": {
      "get": {
        "operationId": "ChatController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          },
          "404": {
            "description": "메시지가 존재하지 않음"
          }
        },
        "summary": "특정 채팅 메시지 조회",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/agent/{agentId}": {
      "get": {
        "operationId": "ChatController_findByAgentId",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessage"
                  }
                }
              }
            }
          }
        },
        "summary": "특정 Agent의 모든 채팅 메시지 조회",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversation": {
      "get": {
        "operationId": "ChatController_findConversation",
        "parameters": [
          {
            "name": "senderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receiverId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessage"
                  }
                }
              }
            }
          }
        },
        "summary": "두 Agent 간의 대화 내용 조회",
        "tags": [
          "Chat"
        ]
      }
    }
  },
  "info": {
    "title": "A2AChat API",
    "description": "Agent 간 상호작용을 위한 API",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Agent의 고유 식별자",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Agent의 이름"
          },
          "description": {
            "type": "string",
            "description": "Agent에 대한 설명"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CreateAgentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Agent의 이름"
          },
          "description": {
            "type": "string",
            "description": "Agent에 대한 설명"
          }
        },
        "required": [
          "name"
        ]
      },
      "Mcp": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "MCP 서버의 URL",
            "format": "uri"
          },
          "token": {
            "type": "string",
            "description": "MCP 서버 인증 토큰"
          },
          "options": {
            "type": "object",
            "description": "추가 설정 옵션",
            "additionalProperties": true
          }
        },
        "required": [
          "url",
          "token"
        ]
      },
      "CreateMcpDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "MCP 서버의 URL",
            "format": "uri"
          },
          "token": {
            "type": "string",
            "description": "MCP 서버 인증 토큰"
          },
          "options": {
            "type": "object",
            "description": "추가 설정 옵션",
            "additionalProperties": true
          }
        },
        "required": [
          "url",
          "token"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "예약 실행의 고유 식별자",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "description": "실행할 Agent의 ID",
            "format": "uuid"
          },
          "cronExpression": {
            "type": "string",
            "description": "실행 주기를 나타내는 Cron 표현식"
          },
          "description": {
            "type": "string",
            "description": "예약 실행에 대한 설명"
          },
          "enabled": {
            "type": "boolean",
            "description": "예약 실행 활성화 여부",
            "default": true
          }
        },
        "required": [
          "id",
          "agentId",
          "cronExpression",
          "enabled"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "agentId": {
            "type": "string",
            "description": "실행할 Agent의 ID",
            "format": "uuid"
          },
          "cronExpression": {
            "type": "string",
            "description": "실행 주기를 나타내는 Cron 표현식"
          },
          "description": {
            "type": "string",
            "description": "예약 실행에 대한 설명"
          },
          "enabled": {
            "type": "boolean",
            "description": "예약 실행 활성화 여부",
            "default": true
          }
        },
        "required": [
          "agentId",
          "cronExpression",
          "enabled"
        ]
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "메시지의 고유 식별자",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "description": "메시지를 보낸 Agent의 ID",
            "format": "uuid"
          },
          "receiverId": {
            "type": "string",
            "description": "메시지를 받는 Agent의 ID",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "메시지 내용"
          },
          "timestamp": {
            "type": "string",
            "description": "메시지 전송 시간",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "senderId",
          "receiverId",
          "content",
          "timestamp"
        ]
      },
      "CreateChatMessageDto": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "string",
            "description": "메시지를 보낸 Agent의 ID",
            "format": "uuid"
          },
          "receiverId": {
            "type": "string",
            "description": "메시지를 받는 Agent의 ID",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "메시지 내용"
          }
        },
        "required": [
          "senderId",
          "receiverId",
          "content"
        ]
      }
    }
  }
}